Correcciones de ABMs y listboxs de vistas
Correcciones en los ABMs para tomar los datos y listas cargadas en las vistas de VerPreguntas y VerPregunta
Tambien agregue un metodo que hacia falta en el HelperABM para obtener la lista de todas las preguntas con sus  listas de respuestas cargadas

+++ Trabajo_Final_Taller2/AccesoADatos/ABMUsuario.cs
@@ -60,7 +60,7 @@ namespace AccesoADatos
                 SqlDataReader reader = command.ExecuteReader();
                 //Utilizo el reader para leer los datos del usuario y cargarlos a user
                 reader.Read();
-                user.IdUsuario = reader.GetInt16(0);
+                user.IdUsuario = reader.GetInt32(0);
                 user.Nombre = reader.GetString(1);
                 user.Apellido = reader.GetString(2);
                 user.PaisOrigen = reader.GetString(3);
